Add Cidr Ignore Filter Option
With the new option it is possible to configure endpoint IP
ranges that should not be used for DNS assignment.

There are cloud providers that provide multiple IPs
for load balancers. For example in OpensStack
the cloud provider can be configured to use Floating IPs
for the load balancers, which are deployed to a local
network (for example the node subnet itself). In such a
scenario the endpoint contains multiple IP addresses on
different networks. But only one should be used for the
DNS entry.

Now it is possible to configure CIDRs that should be
ignored for endpoint IPs. Specifying the CIDR for the
node network (example above) would prevent using those
local IPs for the DNS entry generation.