Add a domain filter option for filtering domains.
The `domain-filter` option is effectively used to filter no domains but hosted
zones. Only zones for a base domain being a (sub)domain of the values specified
for the domain filter are taken into account for the DNS provider, For example
see function

```
func (p *AWSProvider) Zones() (map[string]*route53.HostedZone, error)
```

in [aws.go](https://github.com/kubernetes-incubator/external-dns/blob/master/provider/aws.go).

But another useful scenario (we would require) is to limit the DNS names
handled by the controller to a dedicated base domain assigned for the cluster
(for example: mycluster.acme.com), just reusing a common hosted zone shared
with other clusters or clients.

Therefore it would be nice to be able to specify a `domain-filter` for the DNS
names valid for this cluster. Unfortunately this name is now already used for
the hosted zone filter.

This commit actually uses the option `basedomain-filter` to introduce such
a DNS filter.