Cleanup option
With the new option --cleanup it is possible to remove all DNS entries
managed according to the actual argument settings. This option implies
--once. It should be called after the controller has been stopped.
Otherwise the entries might reappear, before the controller is finished
later on. Access to kubernetes is not required, so it can even be called
after a kubernetes cluster has been destroyed.